migrate utils to path-based `inject()`
