+++ src/utils/inject/css_map.js
@@ -1,0 +1,14 @@
+{
+  const { dataset } = document.currentScript;
+
+  window.tumblr.getCssMap()
+    .then(result => { dataset.result = JSON.stringify(result ?? null); })
+    .catch(exception => {
+      dataset.exception = JSON.stringify({
+        message: exception.message,
+        name: exception.name,
+        stack: exception.stack,
+        ...exception
+      });
+    });
+}
